+++ packages/core/src/core/effect.ts
@@ -77,7 +77,7 @@ class EffectImpl extends ReactiveNode<void> implements EffectObject, DependencyT
       this._notifyCallback = () => scheduler.schedule(this);
     }
 
-    debug.attachDebugInfo(this, 'effect', this.id);
+    debug.attachDebugInfo(this, 'effect', this.id, options.name);
   }
 
   public run(): void {